requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudinary==1.25.0
coverage==5.4
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.7.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
pipenv==2020.11.15
psycopg2-binary==2.8.6
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dotenv==0.17.0
python-http-client==3.3.2
regex==2020.11.13
requests==2.25.1
sendgrid==6.6.0
six==1.15.0
starkbank-ecdsa==1.1.0
stripe==2.56.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==3.4.1
